requirements.txt
asn1crypto==0.24.0
azure-common==1.1.15
azure-nspkg==2.0.0
azure-storage==0.36.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
cycler==0.10.0
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
numpy==1.15.1
Pillow==5.2.0
pycparser==2.19
pyparsing==2.2.1
python-dateutil==2.7.3
python-dotenv==0.9.1
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
